# internal/gen/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Judex API
  version: 1.0.0
  summary: Judex 协作平台公共 API（docs/plans/v1/06 契约源）
  description: |
    所有业务路径前缀 /api/v1（本规范路径不含该前缀，由网关/服务器统一挂载）。

    协议规则：JSON 字段 camelCase；UUID 字符串；时间 RFC3339 UTC；版本为正整数。
    成功 envelope `{data, meta:{requestId,serverTime,eventCursor?}}`；命令额外返回 `affected`。
    错误 envelope `{error:{code,message,details?,retryable},requestId}`；客户端按 code+details 处理，不解析 message。
    分页 `?limit=50&cursor=<opaque>`（1-100）；列表 data 为 `{items,nextCursor}`。
    有副作用的业务命令要求 Idempotency-Key 头（UUID）；已有对象写操作要求 expectedVersion；正式决定要求 reviewId/reviewHash。
    Web 会话使用 Cookie+CSRF（sessionCookie）；CLI 使用受限 Bearer grant（bearerGrant）。
    阶段实现状态由 GET /system capabilities 返回；未实现的 operation 返回 501 而不是假成功。
servers:
  - url: /api/v1
    description: Same-origin (SPA or CLI with explicit base URL)
...
security:
  - sessionCookie: []
  - bearerGrant: []
paths:
...
  /projects/{projectId}/release-reports:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      operationId: listReleaseReports
      summary: 版本发布上报分页
      tags:
        - releases
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: 上报分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ReleaseReport'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createReleaseReport
      summary: 上报真实环境版本事实（不是部署命令）
      tags:
        - releases
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseReportRequest'
      responses:
        '201':
          description: 已登记。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ReleaseReport'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/audit:
    get:
      operationId: listAudit
      summary: 可见业务审计分页（无私人提示词）
      tags:
        - audit
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: objectType
          in: query
          schema:
            type: string
        - name: objectId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 审计分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/AuditEntry'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /me/actions:
    get:
      operationId: listMyActions
      summary: 本人待办（按当前绑定计算）
      tags:
        - me
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: kind
          in: query
          schema:
            type: string
            enum:
              - submit
              - approve
              - receive
              - accept
              - revise
              - reopen
      responses:
        '200':
          description: 待办分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PendingAction'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /projects/{projectId}/tasks/{taskId}/fix-propagations:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: createFixPropagation
      summary: 为 Bug 选择修复传播目标（生成工作提案草稿）
      tags:
        - tasks
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FixPropagationRequest'
      responses:
        '202':
          description: 已生成提案草稿引用。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - proposalId
                    properties:
                      proposalId:
                        type: string
                        format: uuid
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /projects/{projectId}/repositories:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      operationId: listRepositories
      summary: 仓库元数据列表
      tags:
        - repositories
      security:
        - sessionCookie: []
        - bearerGrant: []
      responses:
        '200':
          description: 仓库分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/RepositoryLink'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createRepository
      summary: 登记仓库元数据（manager；不操纵 Git）
      tags:
        - repositories
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryDraftRequest'
      responses:
        '201':
          description: 已登记。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/RepositoryLink'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/repositories/{repositoryId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: repositoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateRepository
      summary: 更新仓库元数据（manager）
      tags:
        - repositories
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryUpdateRequest'
      responses:
        '200':
          description: 更新结果。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/RepositoryLink'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /system:
    get:
      operationId: getSystem
      summary: 服务版本与公开 capability 状态
      tags:
        - system
      security: []
      responses:
        '200':
          description: 服务信息。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/SystemInfo'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  schemas:
...
    SystemInfo:
      type: object
      required:
        - name
        - version
        - protocolVersion
        - schemaRange
      properties:
        name:
          type: string
        version:
          type: string
        commit:
          type: string
        protocolVersion:
          type: string
        schemaRange:
          type: string
        capabilities:
          type: object
          additionalProperties: true
...
    RequestMeta:
      type: object
      required:
        - requestId
        - serverTime
      properties:
        requestId:
          type: string
        serverTime:
          type: string
          format: date-time
        eventCursor:
          type: integer
          format: int64
          description: 当前项目事件序列游标，仅项目资源响应返回。
    AffectedResource:
      type: object
      required:
        - type
        - id
        - version
      properties:
        type:
          type: string
        id:
          type: string
        version:
          type: integer
          minimum: 1
    CommandMeta:
      allOf:
        - $ref: '#/components/schemas/RequestMeta'
        - type: object
          properties:
            affected:
              type: array
              items:
                $ref: '#/components/schemas/AffectedResource'
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 兜底文案；客户端按 code+details 映射，不解析句子。
        details:
          description: 结构化错误详情，按 code 不同而不同。
        retryable:
          type: boolean
          default: false
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNSUPPORTED_SCHEMA
        - INVALID_REFERENCE
        - EMAIL_IN_USE
        - UNAUTHENTICATED
        - SESSION_EXPIRED
        - GRANT_REVOKED
        - FORBIDDEN
        - HUMAN_CONFIRMATION_REQUIRED
        - NOT_FOUND
        - VERSION_CONFLICT
        - REVIEW_STALE
        - DEPENDENCY_CHANGED
        - INVALID_TRANSITION
        - IDEMPOTENCY_CONFLICT
        - SOURCE_VERSION_CONFLICT
        - EVENT_CURSOR_EXPIRED
        - REQUIREMENT_UNMET
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA
        - RATE_LIMITED
        - BUDGET_EXHAUSTED
        - DEPENDENCY_UNAVAILABLE
        - MODEL_UNAVAILABLE
        - SANDBOX_UNAVAILABLE
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          type: string
...
    PendingAction:
      type: object
      required:
        - id
        - kind
        - objectType
        - objectId
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - submit
            - approve
            - receive
            - accept
            - revise
            - reopen
        projectId:
          type: string
          format: uuid
        objectType:
          type: string
          enum:
            - proposal
            - handoff
            - task
            - plan
            - topic
        objectId:
          type: string
          format: uuid
        reviewId:
          type:
            - string
            - 'null'
          format: uuid
        dueAt:
          type:
            - string
            - 'null'
          format: date-time
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RepositoryLink:
      type: object
      required:
        - id
        - displayName
        - url
        - provider
        - defaultBranch
        - archivedAt
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        url:
          type: string
        provider:
          type: string
        defaultBranch:
          type: string
        archivedAt:
          type:
            - string
            - 'null'
          format: date-time
        version:
          type: integer
    RepositoryDraftRequest:
      type: object
      required:
        - displayName
        - url
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        url:
          type: string
        provider:
          type: string
        defaultBranch:
          type: string
          default: main
    RepositoryUpdateRequest:
      type: object
      required:
        - expectedVersion
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        displayName:
          type: string
        url:
          type: string
        provider:
          type: string
        defaultBranch:
          type: string
...
    ReleaseReport:
      type: object
      required:
        - id
        - versionLabel
        - environment
        - status
        - reportedBy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        versionLabel:
          type: string
        environment:
          type: string
        url:
          type:
            - string
            - 'null'
        status:
          type: string
          enum:
            - success
            - partial
            - failed
        repositoryCommits:
          type: array
          items:
            type: object
            required:
              - repositoryId
              - commit
            properties:
              repositoryId:
                type: string
                format: uuid
              branch:
                type: string
              commit:
                type: string
        submissionId:
          type:
            - string
            - 'null'
          format: uuid
        reportedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ReleaseReportRequest:
      type: object
      required:
        - versionLabel
        - environment
        - status
      properties:
        versionLabel:
          type: string
          minLength: 1
          maxLength: 100
        environment:
          type: string
          minLength: 1
          maxLength: 100
        url:
          type: string
        status:
          type: string
          enum:
            - success
            - partial
            - failed
        repositoryCommits:
          type: array
          items:
            type: object
            required:
              - repositoryId
              - commit
            properties:
              repositoryId:
                type: string
                format: uuid
              branch:
                type: string
              commit:
                type: string
        submissionId:
          type: string
          format: uuid
    FixPropagationRequest:
      type: object
      required:
        - targetReleaseRefs
      properties:
        targetReleaseRefs:
          type: array
          minItems: 1
          items:
            type: string
    AuditEntry:
      type: object
      required:
        - id
        - actorType
        - source
        - operation
        - objectType
        - occurredAt
      properties:
        id:
          type: string
          format: uuid
        actorType:
          type: string
          enum:
            - user
            - operator
            - agent
            - system
        actorUserId:
          type:
            - string
            - 'null'
          format: uuid
        identityId:
          type:
            - string
            - 'null'
          format: uuid
        source:
          type: string
          enum:
            - web
            - cli
            - delegate
            - timeout
            - worker
            - operator
            - agent
        operation:
          type: string
        objectType:
          type: string
        objectId:
          type: string
        reviewId:
          type:
            - string
            - 'null'
          format: uuid
        reason:
          type:
            - string
            - 'null'
        occurredAt:
          type: string
          format: date-time
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 分页大小，范围 1-100，默认 50。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    CursorParam:
      name: cursor
      in: query
      description: 不透明分页游标，编码范围与过滤条件。
      schema:
        type: string
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同键同载荷重试返回原结果，同键异载荷返回 409。
      schema:
        type: string
        format: uuid
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: 未认证（401，UNAUTHENTICATED/SESSION_EXPIRED/GRANT_REVOKED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 已认证但无权限（403，FORBIDDEN/HUMAN_CONFIRMATION_REQUIRED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或对非成员隐藏存在性（404）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状态/版本/幂等冲突（409，详见 error.code）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: 依赖暂不可用（503，retryable 明示）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-judex_session
      description: Web 会话 Cookie（本地 HTTP 开发使用 judex_dev_session）。
    bearerGrant:
      type: http
      scheme: bearer
      description: CLI access token（受限 grant，无人工审批能力）。